refactor: move the definition of `Control`s to `control` module
